Introducing features of: 1) ignoring stable sectors, 2) seeking the drive head when retrying the reading of sector.
The goal is skipping the reading of stable sectors to spare time on
physical floppy drives. The other goal is moving the floppy drive head
a bit (seeking) when reading a sector again. It is supported by fdrawsys_dev
but copy command does not use it currently.
At the moment of this commit the stable sector means sector with good crc.
These features are optionals and currently only fdrawsys_dev supports those.
To skip stable sectors the user has to pass skip-stable-sectors parameter to SAMdisk.