Introducing Readstats feature.
Readstats feature provides counting how many times a sector is tried to read
and how many times it is read (either good or bad).

In order to make it backward compatible, a disk_is_read event method is added
to Disk class where readstats of legacy tracks can be fixed. If DemandDisk
subclasses the Disk class then fixing readstats is done in its read method
for each read track. Otherwise if Disk is not DemandDisk then fixing readstats
is done in disk_is_read method.

In order to count readstats more precise:
1) The sector adder (Sector::add) now returns NewData instead of Improved
when the new data has good CRC and the old data has bad CRC, or the new data
has suitable checksum method because it means new data technically.
Improved data means partly matching data and the size of new data is better
than size of old data. In case of NewData the counting restarts from 0.
2) A new merge status Matched is introduced in order to make the repairing
messages more precise and to count read attempt better. Earlier if sector data
matched then Unchanged was returned but as such it did not count read attempt.
Unchanged data means the data is ignored, it does not matter.

To store the readstats in a dsk like file, a new RDSK image file is introduced.
The RDSK is very similar to DSK except:
1) It has a new signature.
2) It has a new ReadStats-Info block at the end of file.
3) Its track size is not limited to 64 KB, instead it is limited to 64 MB.
The Dsk methods are refactored a bit so those could be extended to read and
write Rdsk image file.