Introducing paranoia feature alias multiple good sector data extension.
When a sector is difficult to read, it can happen that the data is read
with good crc but in real the sector is bad. It is because the crc can
match not only the good data but other bad data as well. Thus if the user
wants to be sure the data is good then it must be read multiple times
as good sector.

To activate this feature the user has to pass paranoia parameter to
SAMdisk. In paranoia mode a sector is stable if its data is read at least
stability level times and each data is the same (does not matter if the
data is good or bad). The stability level is 3 by default but the user can
change it by passing stability-level parameter to SAMdisk.