Introduced disk retry and track retry capability when copy command is used.
The user can pass disk-retry and/or track-retry parameters to SAMdisk.
Both can be set as "auto" and in that case the copying is continued while
the destination disk can be improved even with one single bad sector data.

If copying is done in normal mode (that is not in merge or repair mode)
and disk-retry is requested then after the first reading of disk the normal
mode is switched to repair mode.

Note that these retry modes work only if the source disk is not constant
otherwise it would be waste of time merging or repairing the same data
again and again.