Handling "disk too slow" case instead of exiting the application.
Earlier when disk was too slow, diskslow_exception was thrown but not
handled thus the application exited, losing all data read.
Now the diskslow_exception is handled by reopening the floppy drive driver
and reading the source track again. This is tried 5 times and if the drive
is still too slow then the reading of source track is given up and
the copying continues on next track or head.