Bugfix in command line parameter handling, namely the optional debug parameter.
Earlier the user had to specify debug parameter and its number without spaces
between the two. This behaviour comes from getopt which is a bit silly.
For example: --debug1 was OK, --debug 1 was not OK.
Now the same can be specified with or without spaces between the debug
parameter and its number.
For example: --debug1 is OK, --debug 1 is also OK.