Bugfix for positining to correct sector offset when its ID is duplicated and fdrawcmd reads it.
When fdrawcmd reads a sector, it checks if its ID is repeated. In that case
it calculates the sector offset, sets it in FdrawcmdSys in order to read
exactly the requested sector. The calculation is modified, earlier the
firstSectorSeen value was subtracted, now it is added.
Why? The firstSectorSeen value is the index of first sector when it was
found in timed track scan. Normally it is found firstly, so the value of
firstSectorSeen is 0. However it is possible that the first sector is found
after the end of first rotation and then its index is usually the number
of sectors in track (track size) -1. For example assume a track with 10
sectors and the first sector is found at 9th (last) position, its
firstSectorSeen value is 9 but is returned at 0th position by FdrawcmdSys.
Now if we want to position to this first sector then its index position is
(0 + track size (10) + 9) mod track size (10) = 9. By the old calculation
its index position would be (0 + 10 - 9) mod 10 = 1 but the sector is not
there.