Introduced WindowsStub.h in order to build fdrawcmd* on not Windows as much as possible.
Although fdrawcmd driver is Windows only, it is useful if it can be
developed on other operating system as well. That is why the changes
are made. The WindowsStub.h contains all kind of Windows related types,
constants, macros, and functions. The functions are implemented so they
do nothing and return error as if error occured.