Introduced rpm-tolerance-permille commandline parameter.
It is used for checking rpm within tolerance in order to have almost
constant sector time offsets from more track scannings.
This solution was also necessary because there is a bug in fdrawcmd driver
which tries to adjust track time and can fail if can not read all sector
IDAMs (see "Safe fall back on 1st sector" which is not always safe).