Bugfix in DumpTrack method so offsets are display better.
Earlier in relative offset mode if an offset was < than previous offset
then only a red question mark was displayed without a separator space.
Now if the offset is 0 then red N.A. is displayed. If the offset is < than
previous offset then red question mark, minus sign and the offset difference
is displayed. E.g. ?-612 which means that offset is < than previous offset
by 612. In this negative case the previous offset is not updated, so it
does not go backward, it remains the same.