Added normalise_datarate (and its helper convert_offset_by_datarate) methods.
The datarates 250K and 300K are considered interchangeable. In order to
handle them easier one of them must be converted to the other one.
In addition the offset (of sector) also must be converted because it
depends on datarate. All these goals are achieved by the new
normalise_datarate method (of Sector class).