Added byte-tolerance-of-time command line parameter.
The 16 * default value of this new parameter was already coded in
Track::COMPARE_TOLERANCE_BITS. Thus added COMPARE_TOLERANCE_BYTES
constant what is indeed the default value.
This parameter will be used in fdrawcmd driver by passing the parameter
to the driver in CmdTimedMultiScan methoid. This way this parameter
will be used everywhere uniformly.