Using byte_tolerance_of time in Sector.cpp and Track.cpp.
Doing so even that can be answered if sectors from another Track object
contains matching tolerated same sector as this sector considering the
offset as well. The answer also requires the tracklen parameter to match
a sector at the end of track with another sector at the beginning
of track. If tracklen is 0 then this matching is omitted.