Bugfix plus optional paramater of FdrawcmdSys::Seek.
The head parameter is now optional (earlier it was mandatory, before
that it did not exist).
There was a bug which did not cause problem luckily. When the head
parameter was not set in the FD_SEEK_PARAMS structure (e.g. when the
head parameter did not exist) the wrong output size was passed to the
fdrawcmd driver which got the head parameter from the structure which
was auto set as 0 and since this is the default value as well it did not
cause problem.
Solution: when not passing head to the fdrawcmd driver the output size
is decreased by the size of head variable.