Moved PlatformConfig from SAMdisk.h and added missing includes for Windows.
The PlatformConfig.h is moved from SAMdisk.h and it is included
in Platform.h now.
The IncludesByConfig.h is also moved from SAMdisk.h but it is not added
to this commit because probably it will not be necessary. Instead adding
its requested includes on demand.
The includes are added, adjusted for Windows version everywhere.