Refactored FdrawcmdSys in order to print debug messages in debug version.
The Ioctl method had to be wrapped with a IOCTL and RETURN_IOCTL method
instead of another method to avoid processing debug strings in
release version. Since these methods must have optional parameters
(the debug string) and the Ioctl method also had optional
parameters, the parameters of Ioctl method had to be packed in a single
IOCTL_PARAMS structure (else the optional parameters would have been
ambigous). Thus there is a new Ioctl method with the new parameter which
is used in the new IOCTL and RETURN_IOCTL methods but the old Ioctl
method is also kept because it is used by the new method.

The IOCTL macro requires a bool parameter where the returned success is
stored, an ioctl_param structure and optionally a debug message. Its
value is also the bool parameter.
The RETURN_IOCTL macro requires an ioctl_param structure and optionally
a debug message. It is not a value, instead it returns the value as its
name hints.