Improved FdrawcmdSys.CmdReadTrack method.
After doing a lot of tests I found some rules about what FDC requires
and what it returns (I do not have such document). These rules are
built in the CmdReadTrack method which should always receive the amount
of read bytes from the FDC now. As such it is also verified by comparing
to expected amount of bytes.
In addition the size of mem parameter can be 0 now in which case its size
is automatically calculated (it is comfortable).