Added writing Atari ST format feature.
Writing Atari ST instead of raw ensures that the boot sector of Atari ST
filesystem is correct. This method has the capability to restore boot
sector if enough good sectors are available from the FAT and root
directory.
In addition if a sector is bad then the "BADS" text is written at the
beginning of that sector in ths ST file.