Improved sector searching in DiskUtil.RepairTrack.
When a new sector is added to the track as part of repairing track,
its position is searched by finding any ids of following sectors in
the target track. Then the new sector was inserted before the found
sector.
This method is improved in case sector offsets are available. The method
goes backwards from the found sector until it finds another sector
of which offset is <= than the offset of new sector. Then the new sector
is inserted behind that found sector.