Refactored st module and Fat12Filesystem.
The PrepareBootSector is now built into Fat12Filesystem constructor
and the BPB values are written directly into boot sector. If the boot
sector does not exist then it is created. It is a bit of hacking because
the Disk class provides only constant values but it must be modified
to have proper boot sector so the const is removed from returned Track
object.
If a boot sector is created because its data was not available then
the MISS string is written to its beginning (meaning missing sector).