Removed opt_{rpm,rpm_time_tolerance_permille} from fdrawsys_dev.cpp and modified handlong of wrong rpm.
Earlier those were added to customise the rpm checking of tracks.
However those are not necessary by the new handling of tracktime.
In older (<1.0.1.12) driver the tracktime is corrupted rarely. Detecting
it and retrying the scanning is the new solution which also fine tunes
the encoding and data rate detecting.