Format class got ToString, to_string, IsNone, IsSameCylHeadSectorsSize methods.
Earlier it was not possible to check if a Format object is default or not
(more exactly the sectors was checked, and if 0 then Format object is
default, however this assumes that all RegularFormat sets sector > 0).
Thus the RegularFormat::None is introduced and the regular format is
stored in Format object. If it is None then the Format object is default.
The IsSameCylHeadSectorsSize is a helper method to compare the 4 main
format parameters.