Strongly virtualised Disk class and applied some optimisation.
This change was necessary for possible interesting Disk implementations.
Also optimised code where it was necessary. For that a readNC and writeNC
method is added to Disk class which should be called only from Disk
implementators (not only Disk inheritors, thus those are public).
E.g. Disk.write(CylHead, Track) method first wrote the track then
returned the read track. That is it read the track again which might
be good for verifying but here it was not the purpose. So now it
returns the written track and it does not read anyhing in write methods.
This required the writeNC method.