Improved ImageToImage method by using RepairSummaryDisk, the TransferTrack.
The goal was to not lose read tracks from device disk because in case of
weak floppies reading tracks can be a lot of effort. Before copying tracks
the filesystem is detected (if user opted by detect_devfs) on device disk
for which some tracks are read. To keep the best repaied version of those
tracks the RepairSummaryDisk is used.