Improvement in ImageToImage, more effective cache handling in src disk.
In each disk round the cache of src device disk is cleared so the tracks
are forced to read again from the device (useful for detecting stable
sectors). However this affected the reading of filesystem as well so
now only the opted range is cleared.
A helper method clear is added to Disk module.