Bugfix: replaced clear(Range) with clearCache(Range) in Disk module.
It was necessary because clearing tracks resulted in a bug. The copier
knows the track and head amount by the inserted last empty track. If this
track is cleared then the track and head amount are also lost.
Thus now clearing only the cache (m_loaded) flags solves the problem.