Add option to view bit shifted track data
Added --bit-skip option to specify how many bits to skip at the start of a
bitstream when viewing a tracking using the -v option. When specified this
decodes the rest of the track as MFM data bytes with no synchronisation.

The options is useful for viewing sections of data that may not be aligned
to the normal A1 sync bytes, for whatever reason. This includes sections
between a typical write splice at the index and the first sync bytes.